# docs/OpenAPI.yml
openapi: 3.0.0
...
servers:
# Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/AnorZaken7/LoneStarLudo/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/AnorZaken7/LudoMoves/1.0.0
info:
  description: API for the Ludo Lone Star game server.
  version: "2.0.0"
  title: Ludo Lone Star API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /ludo/lobby:
    get:
      operationId: ludoListLobbies
      tags:
        - users
      summary: Lists Ludo lobbies.
      description: TODO
      parameters:
        - name: show
          in: query
          description: |
            Filters lobbies based on how full they are;
            open - Show open only.
            full - Show full only.
            all - Show both open and full.
            penultimate - Show only lobbies with exactly one unoccupied slot.
          schema:
            type: string
            default: open
            enum: [open, full, all, penultimate]
        - name: userId
          in: query
          description: >
            Filter to lobbies where at least one specified user is present
            (e.g. your friends).
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: A JSON array of LobbyListEntry objects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LobbyListEntry'
        '400':
          description: bad input parameter
        '404':
          description: One or more of the specified players does not exists.

    post:
      operationId: ludoCreateLobby
      tags:
        - users
      summary: Creates a new lobby.
      description: Creates a new lobby and joins it as player 0.
      requestBody:
        description: Information required to create a new lobby.
        required: true
        content:
          application/JSON:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CreateLobby'
                - $ref: '#/components/schemas/LoadLobby'
      responses:
        '201': # Created
          description: >
            Returns the id of the newly created game in the Location header.
            The creator of a lobby is always assigned to slot 0 when creating a new game.
        '400':
          description: Player not specified.
        '404':
          description: No such user exists (userId included).
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: SlotCount is out of range or invalid save data.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
        #TODO: some form of auth?
...
  /ludo/board:
    get:
      operationId: ludoGetBoardInfo
      tags:
        - users
      summary: Get a BoardInfo object.
      description: Extrapolates standard board information from a boardLength.
      parameters:
        - name: length
          in: query
          description: Length of the shared competative track around the board.
          schema:
            type: integer
            format: int32
            minimum: 24
            maximum: 80
            default: 40
      responses:
        '200':
          description: A BoardInfo object.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/BoardInfo'
        '400':
          description: Bad length parameter.
  /ludo/board/{gameId}:
    parameters: 
      - $ref: '#/components/parameters/gameIdPath'
    
    get:
      operationId: ludoGetBoardState
      tags:
        - users
      summary: Get current board state.
      description: TODO # <----------------------------------
      responses:
        '200':
          description: OK
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/BoardState'
        '404':
          description: No such game exists.
        '409':
          description: Game is not in the ingame phase.
    
    put:
      tags:
        - players
      summary: Set the BoardState.
      description: Setting the BoardState can only be done while still in lobby.
      operationId: ludoSetBoardState
      requestBody:
        description: The board state (BoardState object containing PieceDistances).
        required: true
        content:
          application/JSON:
            schema:
              $ref: '#/components/schemas/BoardState'
      responses:
        '204':
          description: OK
        '400':
          description: Bad distances parameter.
        '404':
          description: No such game exists.
        '409':
          description: Game has already started (perhaps even ended).
      #TODO: require game auth token + 401 & 403 response
  /ludo/game:
    get:
      operationId: ludoListGames
      tags:
        - users
      summary: Returns a list of games.
      description: Only returns ingame games by default.
      parameters:
        - name: show
          in: query
          description: Filter the game list (ingame, finished, all).
          schema:
            type: string
            enum: [ingame, finished, all]
            default: ingame
        - name: gameId
          in: query
          description: |
            Search for a specific game (implies show=all).
            Note that userId is ignored if gameId is specified.
          schema:
            type: string
        - name: userId
          in: query
          description: |
            Search for all games with this user.
            Note that userId is ignored if gameId is specified.
          schema:
            type: string
      responses:
        '200':
          description: A JSON array of GameListEntry objects.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/GameListEntry'
        '400':
          description: bad input parameter
        '404':
          description: No such game or user exists (only in searches).
  /ludo/game/{gameId}:
    parameters: 
      - $ref: '#/components/parameters/gameIdPath'
    
    get:
      operationId: ludoGetCurrent
      tags:
        - users
      summary: Get current turn, player (slot index), and die roll.
      description: |
        The state of a game (TurnCount, CurrentPlayer, CurrentDieRoll).
        Combine this with board state and lobby data to have a complete save.
      responses:
        '200':
          description: OK
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/TurnSlotDie'
        '404':
          description: No such game exists.
        '409':
          description: Game is not in the ingame phase.
  /ludo/game/{gameId}/{slot}:
    parameters:
      - $ref: '#/components/parameters/gameIdPath'
      - $ref: '#/components/parameters/slotPath'
    # get: decision making info
    # post: PassTurn!
    # delete: concede.
    
    get:
      operationId: ludoGetTurnInfo
      tags:
        - users
      summary: Get a TurnInfo object for the current player (slot index).
      description: The TurnInfo object contains information on the current player's available moves.
      responses:
        '200':
          description: OK
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/TurnInfo'
        '400':
          description: bad input parameter
        '404':
          description: No such game / slot exists.
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Slot does not correspond to the current player. (It's not your turn.)
          content:
            application/JSON:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      operationId: ludoPassTurn
      tags:
        - players
      summary: Ends the turn without moving.
      description: >
        Passes the turn over to the next player; unless the player has a bonus turn ("IsLucky"),
        in which case the current player remains active and only the die is re-rolled.
        Passing is by default only allowed if no move is possible on the current turn.
      responses:
        '204':
          description: OK
        '400':
          description: Bad slot.
        '404':
          description: No such game exists.
        '409':
          description: >
            Rules do not allow passing the turn, or not the active player
            (e.g. game has ended / not started yet).
      #TODO: require game auth token + 401 & 403 response
    
    delete:
      operationId: ludoConcede
      tags:
        - players
      summary: Concede.
      description: |
        Irrevocably hands over control of your pieces to a bot.
        If only one human player remains after a concede, that player is declared the winner and the game ends.
        (wip)
      responses:
        '204':
          description: OK (Also sent if the player has already conceded and / or the game has ended.
        '400':
          description: Bad slot.
        '404':
          description: No such game exists.
        '409':
          description: This player has already conceded and / or the game has ended.
      #TODO: require game auth token + 401 & 403 response
...
components:
  parameters:
...
    gameIdPath:
      name: gameId
      in: path
      required: true
      description: Id of a game.
      schema:
        type: string
    slotPath:
      name: slot
      in: path
      required: true
      description: Game specific slot.
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 3
...
  schemas:
    Error:
      type: object
      parameters:
        code:
          description: An error code.
          type: integer
          format: int32
        desc:
          description: Optional human readable description.
          type: string
    GameState:
      type: string
      enum: [setup, ingame, finished]
      description: The state of a game / lobby.
    LobbyAccess:
      type: string
      enum: [public, unlisted, friendsOnly, inviteOnly, reservations]
      description: |
        Setting that determines the rules for joining a lobby.
        public - game is publicly listed and anyone can join.
        unlisted - anyone with the gameId can join, but game is not publicly listed.
        friendsOnly - game is publicly listed, but only friends can join. (wip)
        inviteOnly - game is not publicly listed and can only be joined by invite. (wip)
        reservations - public game where anyone can join; however slotes are pre-associated
        with playerIds such that if those players join they are automatically granted their pre-associated
        index, vacating any previous occupant from that slot. (wip)
    CreateLobby:
      type: object
      description: Information required to create a new lobby.
      required:
        - userId
      properties:
        userId:
          description: The user creating the lobby.
          type: string
        slots:
          description: Number of slots.
          type: integer
          format: int32
          minimum: 2
          maximum: 4
          default: 4
        access:
          $ref: '#/components/schemas/LobbyAccess'
        reservations:
          description: An array of all slot reservations in this lobby.
          type: array
          items:
            $ref: '#/components/schemas/LobbyReservation'
    LoadLobby:
      type: object
      description: Information required to create a lobby from a saved game.
      required:
        - save
      properties:
        save:
          $ref:'#/components/schemas/GameSave'
        access:
          $ref: '#/components/schemas/LobbyAccess'
        strict:
          type: boolean
          description: >
            Reservations are always generated for the old participants when loading a save.
            This determines if those reservations are strict or relaxed.
          default: false
...
    LobbyReservation:
      type: object
      required:
        - player
        - slot
      properties:
        player:
          type: string
        slot:
          description: Slot index, or -1 if the reservation is slot agnostic.
          type: integer
          format: int32
          minimum: -1
          maximum: 3
        strict:
          type: boolean
          description: >
            If the reservation is strict, then no other player may take this slot.
            Conversely a relaxed reservation behaves like a normal slot, except
            that a player that has a reservation to may evict non-reserved players
            to claim it.
          default: false
    LobbyListEntry:
      type: object
      properties:
        gameId:
          type: string
        access:
          $ref: '#/components/schemas/LobbyAccess'
        slots:
          description: >
            An array of players in slot index order.
            Empty slots are represented by empty array items.
          type: array
          items:
            type: string
        others:
          description: >
            An array of players that are not currently assigned to any slot.
          type: array
          items:
            type: string
        #TODO: add reservation info <-------------------------
...
    GameListEntry:
      type: object
      properties:
        gameId:
          type: string
        state:
          $ref: '#/components/schemas/GameState'
        winner:
          type: string
          description: Winning player (omitted / empty if no winner has been decided).
        slots:
          type: array
          description: A slot array of userIds (or null for empty slots).
          items:
            type: string
        #TODO: add more properties..?
    StartEndPos:  # min-max values derive from min-max of boardLength.
      type: object
      required:
        - startPos
        - endZonePos
      properties:
        startPos:
          type: integer
          description: |
            Absolute board position for where a player places their piece when they take it out of their base.
          format: int32
          minimum: 0
          maximum: 60
        endZonePos:
          type: integer
          description: |
            Absolute board position that corresponds to the first end-zone position of a player.
            Any position larger than or equal to this position corresponds to an endzone position.
          format: int32
          minimum: 24
          maximum: 110
    BoardInfo:  # default values are for a standard sized Ludo board.
      type: object
      required:
        - boardLength
      properties:
        boardLength:
          type: integer
          description: |
            Length of the competative shared track around the board (does not include the safe end-zones leading to the goal).
            Must be a multiple of 8.
          format: int32
          minimum: 24
          maximum: 80
          default: 40
        endZoneLength:
          type: integer
          description: |
            Length of the collision-free end-zones leading to the goal (goal square included). This is usually just boardLength / 8.
            The goal distance - the player relative position / distance a piece must move to reach the goal - is simply boardLength + endZoneLength.
            Thus a piece is in its (safe) end-zone if its distance is greater than boardLength, but smaller than the goal distance.
          format: int32
          minimum: 3
          maximum: 10
          default: 5
        startEndPositions:
          type: array
          description: |
            The absolute start positions, paired with the absolute position of the first square of the endzone, for each player.
          items:
            $ref: '#/components/schemas/StartEndPos'
        #TODO add more properties..?
    PieceDistances:
      type: array
      description: |
        An array of piece distances (for a player's pieces).
        These player relative board positions describe how far a piece has moved.
        To translate them into absolute board positions, the length of the board and the index of the player must also be known.
        Assuming the distance is less than the board length (i.e. if the piece is still on the shared competative track) the formula looks like so;
        position = (distance + slot * boardLength / 4) % boardLength
      items:
        type: integer
        format: int32
        minimum: 0
        maximum: 90 # based on maximum allowed boardLength
    BoardState:
      type: array
      description: An array of piece distance arrays (one for each player, i.e. slot index).
      items:
        $ref: '#/components/schemas/PieceDistances'
    TurnSlotDie:
      type: object
      required:
        - turn
        - player
        - die
      properties:
        turn:
          type: integer
          description: Which turn the game is on.
          format: int32
          minimum: 0
        slot:
          type: integer
          description: Slot index of current player.
          format: int32
          minimum: 0
          maximum: 3
        die:
          type: integer
          description: Die roll for the current player / turn.
          format: int32
          minimum: 1
          maximum: 6
    PlayerSlot:
      type: object
      properties:
        player:
          type: string
        slot:
          description: Slot index, or -1 if not assigned to any slot.
          type: integer
          format: int32
          minimum: -1
          maximum: 3
    LobbySave:
      type: object
      required:
        - players
      properties:
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerSlot'
    GameSave:
      type: object
      required:
        - board
        - current
        - lobby
      properties:
        board:
          $ref: '#/components/schemas/BoardState'
        current:
          $ref: '#/components/schemas/TurnSlotDie'
        lobby:
          $ref: '#/components/schemas/LobbySave'
        #TODO is any data missing to make this a complete game save?
    TurnInfo:
      type: object
      required:
        - pieces
      properties:
        canPass:
          type: boolean
          description: >
            Can the player pass the current turn?
            (Can be omitted if false.)
        isLucky:
          type: boolean
          description: >
            Will the player get a bonus turn after the current one?
            (Can be omitted if false.)
        pieces:
          type: array
          description: A JSON array of PieceInfo objects.
          items:
            $ref: '#/components/schemas/PieceInfo'
    PieceInfo:
      type: object
      required:
        - distance
      properties:
        distance:
          type: integer
          format: int32
          minimum: 0
          description: |
            How far this piece has moved (player-relative board index).
            Zero if it is currently in the players base.
        position:
          type: integer
          format: int32
          minimum: 0
          description: |
            Where this piece is on the board (absolute board index).
            Omitted if the piece isn't on the board, i.e. if it has reached the goal or is in the players base.
            (Distance is non-zero and position is omitted == piece has reached the goal.)
        moved:
          type: integer
          format: int32
          minimum: -1 # if the move puts it at the goal
          description: |
            Where this piece will be on the board if it is moved.
            Omitted if the piece cannot move.
        collision:
          $ref: '#/components/schemas/PlayerPiece'
    PlayerPiece:
          type: object
          required:
            - player
            - piece
          description: |
            A player and piece tuple for describing the content of a target square on the board.
            For example used to describe what a piece would collide with if it were to be moved. (Does not imply that the move in question is legal!)
            Omitted if the target square is empty / there is no collision.
          properties:
            player:
              type: integer
              format: int32
              minimum: 0
              maximum: 3
              description: Zero-based slot index.
            piece:
              type: integer
              format: int32
              minimum: 0
              maximum: 3
              description: Zero-based piece index.